Move gas planet texture loading to own file
